feat: add pathname based conditional rendering to SocialAuthOptions and move to common component directory
