+++ src/components/ui/SocialAuthOptions.jsx
@@ -2,8 +2,12 @@
 import React from "react";
 import Link from "next/link";
 import Image from "next/image"; //
+import { usePathname } from "next/navigation";
 
 export default function SocialAuthOptions() {
+  const pathName = usePathname();
+  const isLoginPage = pathName === "/login";
+  const isRegistrationPage = pathName === "/registration";
   return (
     <>
       <div className="bg-[#E6F2FF] text-secondary-800 rounded-lg px-6 py-4 w-full font-medium leading-[26px]">
@@ -37,17 +41,34 @@ export default function SocialAuthOptions() {
           </div>
         </div>
       </div>
-      <div className="flex gap-1 justify-center items-center">
-        <span className="text-secondary-800 text-sm font-medium leading-[24px]">
-          이미 회원이신가요?
-        </span>
-        <Link
-          href="/login"
-          className="text-primary underline text-sm font-medium leading-[24px]"
-        >
-          로그인
-        </Link>
-      </div>
+
+      {isRegistrationPage && (
+        <div className="flex gap-1 justify-center items-center">
+          <span className="text-secondary-800 text-sm font-medium leading-[24px]">
+            이미 회원이신가요?
+          </span>
+          <Link
+            href="/login"
+            className="text-primary underline text-sm font-medium leading-[24px]"
+          >
+            로그인
+          </Link>
+        </div>
+      )}
+
+      {isLoginPage && (
+        <div className="flex gap-1 justify-center items-center">
+          <span className="text-secondary-800 text-sm font-medium leading-[24px]">
+            판다마켓이 처음이신가요?
+          </span>
+          <Link
+            href="/registration"
+            className="text-primary underline text-sm font-medium leading-[24px]"
+          >
+            회원가입
+          </Link>
+        </div>
+      )}
     </>
   );
 }
